+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Changes
 
 * [BREAKING][param][rust] `NodeRpcClient::get_block_by_number()` now takes an `include_proof: bool` parameter to control whether the block proof is included in the response. ([#TBD](https://github.com/0xMiden/miden-client/pull/TBD))
-* [BREAKING][param][rust] `NodeRpcClient::sync_chain_mmr()` replaced `block_to: Option<BlockNumber>` with `chain_tip: ChainTip` to match the RPC definition. Use `ChainTip::Committed` for previous default behavior (`None`). ([#TBD](https://github.com/0xMiden/miden-client/pull/TBD))
+* [BREAKING][param][rust] `NodeRpcClient::sync_chain_mmr()` replaced `block_to: Option<BlockNumber>` with `upper_bound: SyncTarget` to match the RPC definition. Use `SyncTarget::CommittedChainTip` for previous default behavior (`None`), or `SyncTarget::BlockNumber(num)` for a specific block number. ([#TBD](https://github.com/0xMiden/miden-client/pull/TBD))
 
 ### Enhancements